requirements.txt
pytest==6.2.5
requests==2.28.1
pydantic==1.9.0
allure-pytest==2.9.45
faker==11.3.0
pytest-xdist==2.5.0
pyhamcrest==2.0.4